+++ instrumentation/action_pack/test/test_helpers/app_config.rb
@@ -28,8 +28,6 @@ module AppConfig
     new_app.config.filter_parameters = [:param_to_be_filtered]
 
     case Rails.version
-    when /^6\.0/
-      apply_rails_6_0_configs(new_app)
     when /^6\.1/
       apply_rails_6_1_configs(new_app)
     when /^7\./
@@ -50,9 +48,7 @@ module AppConfig
   private
 
   def remove_rack_middleware(application)
-    application.middleware.delete(
-      OpenTelemetry::Instrumentation::Rack::Middlewares::TracerMiddleware
-    )
+    application.middleware.delete(Rack::Events)
   end
 
   def add_exceptions_app(application)
@@ -71,13 +67,6 @@ module AppConfig
       ActionDispatch::DebugExceptions,
       RedirectMiddleware
     )
-  end
-
-  def apply_rails_6_0_configs(application)
-    # Required in Rails 6
-    application.config.hosts << 'example.org'
-    # Creates a lot of deprecation warnings on subsequent app initializations if not explicitly set.
-    application.config.action_view.finalize_compiled_template_methods = ActionView::Railtie::NULL_OPTION
   end
 
   def apply_rails_6_1_configs(application)