+++ instrumentation/excon/lib/opentelemetry/instrumentation/excon/instrumentation.rb
@@ -4,15 +4,20 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+require_relative '../concerns/untraced_hosts'
+
 module OpenTelemetry
   module Instrumentation
     module Excon
       # The Instrumentation class contains logic to detect and install the Excon
       # instrumentation
       class Instrumentation < OpenTelemetry::Instrumentation::Base
+        include OpenTelemetry::Instrumentation::Concerns::UntracedHosts
+
         install do |_config|
           require_dependencies
           add_middleware
+          patch
         end
 
         present do
@@ -25,11 +30,15 @@ module OpenTelemetry
 
         def require_dependencies
           require_relative 'middlewares/tracer_middleware'
+          require_relative 'patches/socket'
         end
 
         def add_middleware
-          ::Excon.defaults[:middlewares] =
-            Middlewares::TracerMiddleware.around_default_stack
+          ::Excon.defaults[:middlewares] = Middlewares::TracerMiddleware.around_default_stack
+        end
+
+        def patch
+          ::Excon::Socket.prepend(Patches::Socket)
         end
       end
     end