+++ instrumentation/ruby_kafka/lib/opentelemetry/instrumentation/ruby_kafka/instrumentation.rb
@@ -22,7 +22,7 @@ module OpenTelemetry
         end
 
         compatible do
-          (!gem_version.nil? && gem_version >= MINIMUM_VERSION)
+          !gem_version.nil? && gem_version >= MINIMUM_VERSION
         end
 
         private