+++ instrumentation/base/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History: opentelemetry-instrumentation-base
 
+### v0.22.3 / 2023-11-23
+
+* CHANGED: Applied Rubocop Performance Recommendations [#727](https://github.com/open-telemetry/opentelemetry-ruby-contrib/pull/727)
+
+### v0.22.2 / 2023-08-03
+
+* FIXED: Remove inline linter rules
+
 ### v0.22.1 / 2023-06-02
 
 * feat: make config available to compatible blocks #453