+++ instrumentation/bunny/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History: opentelemetry-instrumentation-bunny
 
+### v0.21.2 / 2024-02-08
+
+* FIXED: Remove disable directives leftover from older version of Rubocop
+
+### v0.21.1 / 2023-09-27
+
+* FIXED: Headers property needs to exist
+
+### v0.21.0 / 2023-09-07
+
+* FIXED: Align messaging instrumentation operation names
+
 ### v0.20.1 / 2023-06-05
 
 * FIXED: Base config options